pyproject.toml
[project]
name = "66degrees"
version = "0.1.0"
description = "Reusable Kaggle dataset pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kaggle>=2.0.0",
]

[project.optional-dependencies]
gcs = ["google-cloud-storage>=2.0"]
s3  = ["boto3>=1.34"]
all = ["google-cloud-storage>=2.0", "boto3>=1.34"]

[project.scripts]
pipeline = "main:main"
